Schwinger.jl

Schwinger — Module

A Julia package for the Hamiltonian lattice Schwinger model.

Module Index

Schwinger.EDHamiltonian

Schwinger.MPOHamiltonian

Schwinger.SchwingerBasisState

Schwinger.SchwingerEDState

Schwinger.SchwingerLattice

Schwinger.SchwingerMPS

Schwinger.EDWilsonLoop

Schwinger.MPOWilsonLoop

Schwinger.basis

Schwinger.charges

Schwinger.electricfields

Schwinger.energy

Schwinger.energy

Schwinger.energy

Schwinger.energygap

Schwinger.entanglements

Schwinger.expectation

Schwinger.expectation

Schwinger.groundstate

Schwinger.loweststates

Schwinger.occupations

Schwinger.occupations

Schwinger.occupations

Schwinger.scalarvev

Detailed API

Schwinger.EDHamiltonian — Method

EDHamiltonian(lattice) Computes the Hamiltonian for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.MPOHamiltonian — Method

MPOhamiltonian(lattice)

Computes the MPO Hamiltonian for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.SchwingerBasisState — Type

SchwingerBasisState{N,F}(occupations, L0)

A Schwinger model basis state.

Schwinger.SchwingerEDState — Type

SchwingerEDState{N,F}(hamiltonian, coeffs)

A Schwinger model state represented as a linear combination of basis states.

Schwinger.SchwingerLattice — Type

SchwingerLattice(N, F; periodic=false, L_max=0)

Constructs a SchwingerLattice for the Schwinger model.

Arguments

N::Int: Number of sites.

F::Int: Number of flavors.

periodic::Bool=false: Whether the lattice is periodic.

L_max::Int=0: Maximum absolute value for L_0 when periodic.

Returns

A SchwingerLattice object.

Schwinger.SchwingerMPS — Type

SchwingerMPSState{N,F}(hamiltonian, psi)

A Schwinger model MPS.

Schwinger.EDWilsonLoop — Method

wilsonloop(hamiltonian, conjugate = false)

Returns the spatial Wilson loop operator for lattice.

Arguments

lattice::SchwingerLattice: lattice.

conjugate::Bool: Conjugation of the Wilson loop.

Schwinger.MPOWilsonLoop — Method

wilsonloop(hamiltonian, conjugate = false)

Returns the spatial Wilson loop operator for lattice.

Arguments

lattice::SchwingerLattice: lattice.

conjugate::Bool: Conjugation of the Wilson loop.

Schwinger.basis — Method

basis(N, F; L_max)

Returns a basis of Schwinger model states.

Arguments

N::Int: number of sites.

F::Int: number of flavors.

L_max::Int: maximum value of |L₀|.

Schwinger.charges — Method

charges(state)

Return a list of the expectations of Q operators on each site and for each known eigenstate.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.electricfields — Method

electricfields(state)

Return a list of the expectations of (L + θ/2π) operators on each link.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerEDState: Schwinger model state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.energygap — Method

energygap(hamiltonian) Returns the energy difference between the lowest two states of the Hamiltonian.

Arguments

hamiltonian::SchwingerHamiltonian: Schwinger model Hamiltonian.

Schwinger.entanglements — Method

entanglements(state)

Return a list of the entanglement entropies for each bisection of the lattice.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.expectation — Method

expectation(op, state)

Return the expectation value of the operator op in state.

Arguments

op::EDOperator{N,F}`: operator.

state::SchwingerEDState: state.

Schwinger.expectation — Method

expectation(op, state)

Return the expectation value of the operator op in state.

Arguments

op::MPOOperator{N,F}`: operator.

state::SchwingerMPS: state.

Schwinger.groundstate — Method

groundstate(hamiltonian) Returns the ground state of the Schwinger model Hamiltonian.

Arguments

hamiltonian::SchwingerHamiltonian: Schwinger model Hamiltonian.

Schwinger.loweststates — Method

lowest_states(hamiltonian, nstates)

Returns the lowest few eigenstates of the Schwinger model Hamiltonian.

Arguments

hamiltonian::MPOHamiltonian: Schwinger model Hamiltonian.

nstates::Int:: number of states to determine.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerMPS{N,F}: Schwinger model state.

Schwinger.scalarvev — Method

scalarvev(state)

Return the VEV of the scalar condensate L⁻¹ ∑ (-1)ⁿ χ†ₙχₙ

Arguments

state::SchwingerState: Schwinger model state.