Schwinger.jl

Schwinger — Module

A Julia package for the Hamiltonian lattice Schwinger model.

Module Index

Schwinger.EDHamiltonian

Schwinger.MPOHamiltonian

Schwinger.SchwingerBasisState

Schwinger.SchwingerEDState

Schwinger.SchwingerLattice

Schwinger.charges

Schwinger.electricfields

Schwinger.energy

Schwinger.energy

Schwinger.energy

Schwinger.entanglements

Schwinger.occupations

Schwinger.occupations

Schwinger.occupations

Detailed API

Schwinger.EDHamiltonian — Method

EDHamiltonian(lattice) Computes the Hamiltonian for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.MPOHamiltonian — Method

MPOhamiltonian(lattice)

Computes the MPO Hamiltonian for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.SchwingerBasisState — Type

SchwingerBasisState{N,F}(occupations, L0)

A Schwinger model basis state.

Schwinger.SchwingerEDState — Type

SchwingerEDState{N,F}(hamiltonian, coeffs)

A Schwinger model state represented as a linear combination of basis states.

Schwinger.SchwingerLattice — Type

SchwingerLattice(N, F; periodic=false, L_max=0)

Constructs a SchwingerLattice for the Schwinger model.

Arguments

N::Int: Number of sites.

F::Int: Number of flavors.

periodic::Bool=false: Whether the lattice is periodic.

L_max::Int=0: Maximum absolute value for L_0 when periodic.

Returns

A SchwingerLattice object.

Schwinger.charges — Method

charges(state)

Return a list of the expectations of Q operators on each site and for each known eigenstate.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.electricfields — Method

electricfields(state)

Return a list of the expectations of (L + θ/2π) operators on each link.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerEDState: Schwinger model state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.entanglements — Method

entanglements(state)

Return a list of the entanglement entropies for each bisection of the lattice.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerMPS{N,F}: Schwinger model state.