Schwinger.jl

Schwinger — Module

A Julia package for the Hamiltonian lattice Schwinger model.

Module Index

Schwinger.SchwingerBasisState

Schwinger.SchwingerEDState

Schwinger.SchwingerLattice

Schwinger.SchwingerMPS

Schwinger.EDGaugeKinetic

Schwinger.EDHamiltonian

Schwinger.EDHopping

Schwinger.EDHoppingMass

Schwinger.EDMass

Schwinger.EDWilsonLine

Schwinger.EDWilsonLoop

Schwinger.L₀

Schwinger.MPOHamiltonian

Schwinger.MPOWilsonLine

Schwinger.MPOWilsonLoop

Schwinger.act

Schwinger.act

Schwinger.charges

Schwinger.electricfields

Schwinger.energy

Schwinger.energy

Schwinger.energygap

Schwinger.entanglements

Schwinger.expectation

Schwinger.expectation

Schwinger.groundstate

Schwinger.loweststates

Schwinger.loweststates

Schwinger.occupations

Schwinger.occupations

Schwinger.occupations

Schwinger.pseudoscalarvev

Schwinger.scalarvev

Schwinger.schwingerbasis

Detailed API

Schwinger.SchwingerBasisState — Type

SchwingerBasisState{N,F}(occupations, L0)

A Schwinger model basis state.

Schwinger.SchwingerEDState — Type

SchwingerEDState{N,F}(hamiltonian, coeffs)

A Schwinger model state represented as a linear combination of basis states.

Schwinger.SchwingerLattice — Type

SchwingerLattice(N, F; periodic=false, L_max=0)

Constructs a SchwingerLattice for the Schwinger model.

Arguments

N::Int: Number of sites.

F::Int: Number of flavors.

periodic::Bool=false: Whether the lattice is periodic.

L_max::Int=0: Maximum absolute value for L_0 when periodic.

Returns

A SchwingerLattice object.

Schwinger.SchwingerMPS — Type

SchwingerMPSState{N,F}(hamiltonian, psi)

A Schwinger model MPS.

LinearAlgebra.dot — Method

dot(psi1::SchwingerEDState, psi2::SchwingerEDState)

Return the inner product of two Schwinger model states.

Arguments

psi1::SchwingerEDState: state.

psi2::SchwingerEDState: state.

LinearAlgebra.dot — Method

dot(psi1::SchwingerMPS, psi2::SchwingerMPS)

Return the inner product of two Schwinger model states.

Arguments

psi1::SchwingerMPS: state.

psi2::SchwingerMPS: state.

Schwinger.EDGaugeKinetic — Method

EDGaugeKinetic(lattice) Computes the gauge kinetic operator ∑(Lₙ+θ/2π)² for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.EDHamiltonian — Method

EDHamiltonian(lattice) Computes the Hamiltonian for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.EDHopping — Method

EDHopping(lattice) Computes the hopping term -i ∑(χ†ₙ χₙ₊₁ - χ†ₙ₊₁ χₙ) for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.EDHoppingMass — Method

EDHoppingMass(lattice) Computes the hopping-type mass term i/2 ∑(-1)^n (χ†ₙ₊₁ χₙ + χ†ₙ₋₁ χₙ) for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.EDMass — Method

EDMass(lattice) Computes the mass operator ∑ (-1)ⁿ χ†ₙχₙ for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.EDWilsonLine — Method

EDWilsonLine(lattice, conjugate = false, flavor = 1, start = 1, finish = N)

Returns the spatial Wilson line operator for lattice.

Arguments

lattice::SchwingerLattice: lattice.

conjugate::Bool: Conjugation of the Wilson line.

flavor::Int: Flavor of the Wilson line.

start::Int: Starting site of the Wilson line.

finish::Int: Finishing site of the Wilson line.

Schwinger.EDWilsonLoop — Method

wilsonloop(hamiltonian, conjugate = false)

Returns the spatial Wilson loop operator for lattice.

Arguments

lattice::SchwingerLattice: lattice.

conjugate::Bool: Conjugation of the Wilson loop.

Schwinger.L₀ — Method

L₀(state)

Return the expectation value of L₀.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.MPOHamiltonian — Method

MPOHamiltonian(lattice)

Computes the MPO Hamiltonian for the Schwinger model.

Arguments

lattice::SchwingerLattice: Schwinger model lattice.

Schwinger.MPOWilsonLine — Method

MPOWilsonLine(lattice, conjugate = false, flavor = 1, start = 1, finish = N)

Returns the spatial Wilson line operator for lattice.

Arguments

lattice::SchwingerLattice: lattice.

conjugate::Bool: Conjugation of the Wilson line.

flavor::Int: Flavor of the Wilson line.

start::Int: Starting site of the Wilson line.

finish::Int: Finishing site of the Wilson line.

Schwinger.MPOWilsonLoop — Method

wilsonloop(hamiltonian, conjugate = false)

Returns the spatial Wilson loop operator for lattice.

Arguments

lattice::SchwingerLattice: lattice.

conjugate::Bool: Conjugation of the Wilson loop.

Schwinger.act — Method

act(op, state)

Apply the operator op to the state state.

Arguments

op::EDOperator{N,F}: operator.

state::SchwingerEDState{N,F}: state.

Schwinger.act — Method

act(op, state)

Apply the operator op to the state state.

Arguments

op::MPOOperator{N,F}: operator.

state::SchwingerMPS{N,F}: state.

Schwinger.charges — Method

charges(state)

Return a list of the expectations of Q operators on each site and for each known eigenstate.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.electricfields — Method

electricfields(state)

Return a list of the expectations of (L + θ/2π) operators on each link.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.energy — Method

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerEDState: Schwinger model state.

Schwinger.energygap — Method

energygap(hamiltonian) Returns the energy difference between the lowest two states of the Hamiltonian.

Arguments

hamiltonian::SchwingerOperator: Schwinger model Hamiltonian.

Schwinger.entanglements — Method

entanglements(state)

Return a list of the entanglement entropies for each bisection of the lattice.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.expectation — Method

expectation(op, state)

Return the expectation value of the operator op in state.

Arguments

op::EDOperator{N,F}`: operator.

state::SchwingerEDState: state.

Schwinger.expectation — Method

expectation(op, state)

Return the expectation value of the operator op in state.

Arguments

op::MPOOperator{N,F}`: operator.

state::SchwingerMPS: state.

Schwinger.groundstate — Method

groundstate(hamiltonian) Returns the ground state of the Schwinger model Hamiltonian.

Arguments

hamiltonian::SchwingerOperator: Schwinger model Hamiltonian.

Schwinger.loweststates — Method

lowest_states(hamiltonian, nstates)

Returns the lowest few eigenstates of the Schwinger model Hamiltonian.

Arguments

hamiltonian::EDOperator: Schwinger model Hamiltonian.

nstates::Int:: number of states to determine.

Schwinger.loweststates — Method

lowest_states(hamiltonian, nstates)

Returns the lowest few eigenstates of the Schwinger model Hamiltonian.

Arguments

hamiltonian::MPOOperator: Schwinger model Hamiltonian.

nstates::Int:: number of states to determine.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.occupations — Method

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerMPS{N,F}: Schwinger model state.

Schwinger.pseudoscalarvev — Method

pseudoscalarvev(state)

Return the VEV of the pseudoscalar condensate L⁻¹ ∑ (-1)ⁿ χ†ₙχₙ

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.scalarvev — Method

scalarvev(state)

Return the VEV of the scalar condensate L⁻¹ ∑ (-1)ⁿ χ†ₙχₙ

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.schwingerbasis — Method

schwingerbasis(lattice; L_max, q, universe)

Returns a basis of Schwinger model states.

Arguments

lattice::SchwingerLattice

L_max::Int

q::Int

universe::Int