States

Lattice states in Schwinger.jl are represented by the abstract type SchwingerState, with three descendants:

SchwingerBasisState: a state specified by the eigenvalues of occupation operators $\chi^\dagger_{n,\alpha}\chi_{n,\alpha}$ and $L_0$

SchwingerEDState: a linear combination of SchwingerBasisStates

SchwingerMPS: a matrix product state, stored as an MPS object using ITensorMPS.jl

Given a state, we can find the expectation values of the occupation operators and electric field operators:

using Schwinger
lat = SchwingerLattice{6,1}(a = 10) # towards the lattice strong coupling limit ga -> infty
gs = groundstate(EDHamiltonian(lat))

occupations(gs), electricfields(gs)

([0.9996009569748598; 0.0007976086592019031; … ; 0.999202391340798; 0.0003990430251401028;;], [-0.00039904302514015555; 0.0003985656340617476; … ; -0.0003990430251402126; -1.0977546996415732e-16;;])

We can also evaluate the entanglement entropies of each bisection of the lattice:

using Schwinger
lat = SchwingerLattice{20,1}()
gs = groundstate(MPOHamiltonian(lat))
entanglements(gs)

19-element Vector{Float64}:
 0.5892661594332941
 0.39444995102389974
 0.5168663014805427
 0.4537304990970738
 0.4899161466432022
 0.4701048867712392
 0.4812980874019118
 0.4748933776117984
 0.47884978094410363
 0.4759380630855685
 0.4788497809439815
 0.47489337761173883
 0.4812980874018509
 0.4701048867711649
 0.4899161466431351
 0.45373049909698887
 0.5168663014802909
 0.39444995102369124
 0.5892661594332931

Several other useful functions are detailed below.

Schwinger.SchwingerState — Type

SchwingerState{N,F}

Abstract type for Schwinger model states.

Schwinger.SchwingerBasisState — Type

SchwingerBasisState{N,F}(occupations, L0)

A Schwinger model basis state.

Schwinger.SchwingerEDState — Type

SchwingerEDState{N,F}(hamiltonian, coeffs)

A Schwinger model state represented as a linear combination of basis states.

Schwinger.SchwingerMPS — Type

SchwingerMPSState{N,F}(hamiltonian, psi)

A Schwinger model MPS.

Schwinger.occupation — Function

occupation(state, site)

Return the expectations of χ†χ operators of each flavor on a given site.

Arguments

state::SchwingerMPS{N,F}: Schwinger model state.

site::Int: the lattice site.

occupation(state, site)

Return the expectations of χ†χ operators of each flavor on a given site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

site::Int: the lattice site.

occupation(state, site)

Return the expectations of χ†χ operators of each flavor on a given site.

Arguments

state::SchwingerEDState: Schwinger model basis state.

site::Int: the lattice site.

Schwinger.occupations — Function

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerMPS{N,F}: Schwinger model state.

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerEDState: Schwinger model basis state.

Schwinger.charge — Function

charge(state, site)

Return the expectation value of the charge operator on site site.

Arguments

state::SchwingerState: Schwinger model state.

site::Int: site.

Schwinger.charges — Function

charges(state)

Return a list of the expectations of Q operators on each site and for each known eigenstate.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.electricfield — Function

electricfield(state, link)

Return the expectation of (L + θ/2π) on the link link.

Arguments

state::SchwingerState: Schwinger model state.

link::Int: link.

Schwinger.electricfields — Function

electricfields(state)

Return a list of the expectations of (L + θ/2π) operators on each link.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.entanglement — Function

entanglement(state, bisection)

Return the von Neumann entanglement entropy -tr(ρₐ log(ρₐ)), where a is the subsystem of sites 1..bisection

Arguments

state::SchwingerEDState: Schwinger model state.

bisection::Int: bisection index.

entanglement(state, bisection)

Return the von Neumann entanglement entropy -tr(ρₐ log(ρₐ)), where a is the subsystem of sites 1..bisection

Arguments

state::SchwingerMPS: Schwinger model state.

bisection::Int: bisection index.

Schwinger.entanglements — Function

entanglements(state)

Return a list of the von Neumann entanglement entropies for each bisection of the lattice.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.energy — Function

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerEDState: Schwinger model state.

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.L₀ — Function

L₀(state)

Return the expectation value of L₀.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.scalar — Function

scalar(state)

Return the expectation value of the scalar condensate, ⟨H_mass⟩/L.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.scalardensity — Function

scalardensity(state, site)

Return the scalar density at site site.

Arguments

state::SchwingerState: Schwinger model state.

site::Int: site.

Schwinger.scalardensities — Function

scalardensities(state)

Return the list of scalar densities of state on sites 1 through N.

Arguments

state::SchwingerState: Schwinger model state.

site::Int: site.

Schwinger.pseudoscalar — Function

pseudoscalar(state)

Return the expectation value of the pseudoscalar condensate, ⟨H_hoppingmass⟩/L.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.pseudoscalardensity — Function

pseudoscalardensity(state, n)

Return the pseudoscalar density at site n.

Arguments

state::SchwingerState: Schwinger model state.

n::Int: site.

Schwinger.pseudoscalardensities — Function

pseudoscalardensities(state)

Return the list of pseudoscalar densities of state on sites 1 through N.

Arguments

state::SchwingerState: Schwinger model state.

site::Int: site.