States

Lattice states in Schwinger.jl are represented by the abstract type SchwingerState, with three descendants:

SchwingerBasisState: a state specified by the eigenvalues of occupation operators $\chi^\dagger_{n,\alpha}\chi_{n,\alpha}$ and $L_0$

SchwingerEDState: a linear combination of SchwingerBasisStates

SchwingerMPS: a matrix product state, stored as an MPS object using ITensorMPS.jl

Given a state, we can find the expectation values of the occupation operators and electric field operators:

using Schwinger
lat = SchwingerLattice{6,1}(a = 10) # towards the lattice strong coupling limit ga -> infty
gs = groundstate(EDHamiltonian(lat))

occupations(gs), electricfields(gs)

([0.9996009569748598; 0.0007976086592019031; … ; 0.999202391340798; 0.0003990430251401028;;], [-0.00039904302514015555; 0.0003985656340617476; … ; -0.0003990430251402126; -1.0977546996415732e-16;;])

We can also evaluate the entanglement entropies of each bisection of the lattice:

using Schwinger
lat = SchwingerLattice{20,1}()
gs = groundstate(MPOHamiltonian(lat))
entanglements(gs)

19-element Vector{Float64}:
 0.5892661594332932
 0.39444995102389996
 0.5168663014805462
 0.4537304990970813
 0.4899161466432008
 0.4701048867712738
 0.4812980874018529
 0.47489337761187705
 0.47884978094400155
 0.4759380630856635
 0.47884978094388997
 0.47489337761180844
 0.48129808740180685
 0.47010488677119827
 0.48991614664311556
 0.45373049909699925
 0.5168663014802842
 0.39444995102369274
 0.5892661594332914

Several other useful functions are detailed below.

Schwinger.SchwingerState — Type

SchwingerState{N,F}

Abstract type for Schwinger model states.

Schwinger.SchwingerBasisState — Type

SchwingerBasisState{N,F}(occupations, L0)

A Schwinger model basis state.

Schwinger.SchwingerEDState — Type

SchwingerEDState{N,F}(hamiltonian, coeffs)

A Schwinger model state represented as a linear combination of basis states.

Schwinger.SchwingerMPS — Type

SchwingerMPSState{N,F}(hamiltonian, psi)

A Schwinger model MPS.

Schwinger.occupations — Function

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerMPS{N,F}: Schwinger model state.

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

occupations(state)

Return an NxF matrix of the expectations of χ†χ operators on each site.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.charges — Function

charges(state)

Return a list of the expectations of Q operators on each site and for each known eigenstate.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.electricfields — Function

electricfields(state)

Return a list of the expectations of (L + θ/2π) operators on each link.

Arguments

state::SchwingerMPS: Schwinger model state.

Schwinger.entanglements — Function

entanglements(state)

Return a list of the von Neumann entanglement entropies for each bisection of the lattice.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.energy — Function

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerEDState: Schwinger model state.

energy(state)

Return the expectation value of the Hamiltonian.

Arguments

state::SchwingerBasisState: Schwinger model basis state.

Schwinger.L₀ — Function

L₀(state)

Return the expectation value of L₀.

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.scalarvev — Function

scalarvev(state)

Return the VEV of the scalar condensate L⁻¹ ∑ (-1)ⁿ χ†ₙχₙ

Arguments

state::SchwingerState: Schwinger model state.

Schwinger.pseudoscalarvev — Function

pseudoscalarvev(state)

Return the VEV of the pseudoscalar condensate L⁻¹ ∑ (-1)ⁿ χ†ₙχₙ

Arguments

state::SchwingerState: Schwinger model state.